feat: add a toggle button to the headspace app
This commit adds a toggle button to the headspace app.

The button is used to start and stop the animation.

The animation is now only running when the button is toggled.

This commit also fixes a bug where the animation was not running when the
button was toggled.

The commit also fixes a bug where the animation was running when the button
was not toggled.