feat: add useComputedValue hook
This hook allows us to create a value that is computed from other values.

This is useful for when we want to create a value that is dependent on other
values, but we don't want to re-render the component every time the value
changes.

This is useful for when we want to create a value that is dependent on other
values, but we don't want to re-render the component every time the value
changes.

This is useful for when we want to create a value that is dependent on other
values, but we don't want to re-render the component every time the value
changes.